Make some keyboard shortcuts generally available in plane mode again
